Allow to set bucket location to avoid get_bucket_location request
As noted in
https://github.com/mattheworiordan/capybara-screenshot/pull/239#issuecomment-449782254,
it is currently required to fetch the bucket location to form the S3
URL which now requires the bucket name to be present.

By optionally accepting a `bucket_location` parameter in the s3
configuration, we can save this request and also avoid having to provide
the `s3:GetBucketLocation` permission on the bucket.

The README is extended to reflect this change